Split ObjectStore and related out of restore.py
pylint was rightfully complaining about the excessive line count.
