+++ pghoard/preservation_request.py
@@ -21,7 +21,7 @@ def is_basebackup_preserved(basebackup_entry: Mapping[str, Any], now: datetime.d
 
 
 def parse_preservation_requests(preservation_requests: Sequence[Mapping[str, Any]], ) -> Mapping[str, datetime.datetime]:
-    backups_to_preserve = {}
+    backups_to_preserve: dict[str, datetime.datetime] = {}
     for preservation_request in preservation_requests:
         backup_name = preservation_request["metadata"]["preserve-backup"]
         preserve_until = dates.parse_timestamp(preservation_request["metadata"]["preserve-until"])
